refactor: move the per-IP rate limiter into mulecommon
The External Connection password exchange needs the same brute-force
protection amuleapi's HTTP login already has, but amuled links
mulecommon and not the webapi library, so the class has to live lower
down.

Pure move: the sliding-window logic is byte-identical, and AuthTest --
which constructs the limiter directly and drives it with an injected
clock -- passes unchanged, which is what makes this a refactor rather
than a rewrite.

Only the framing changed. The old comments described the caller
emitting HTTP 429 with Retry-After, which means nothing to a protocol
that has neither; the contract is now stated in terms of refusing
before the credential compare and reporting how long to wait.

`webapi::CRateLimiter` still resolves, so the dispatcher's two members
and every existing call site are untouched.
