+++ src/webapi/Auth.cpp
@@ -95,70 +95,6 @@ void CRevocationSet::GcExpired() const
 	}
 }
 
-// ---------- CRateLimiter -----------------------------------------
-
-CRateLimiter::Decision CRateLimiter::Check(const std::string &ip)
-{
-	std::lock_guard<std::mutex> lock(m_mu);
-	const std::time_t now = m_clock();
-	auto it = m_buckets.find(ip);
-	if (it == m_buckets.end())
-		return Decision{};
-
-	Bucket &b = it->second;
-	if (b.lockout_until > now) {
-		Decision d;
-		d.locked_out = true;
-		d.retry_after_seconds = b.lockout_until - now;
-		return d;
-	}
-	// Lockout window expired — wipe the bucket so a stale lockout
-	// can't accidentally fire on the next Check after a long quiet
-	// period.
-	if (b.lockout_until != 0 && b.lockout_until <= now) {
-		b.lockout_until = 0;
-		b.failures.clear();
-	}
-	// Mirror NoteFailure's per-stamp expiry so Check is self-
-	// consistent. Otherwise stale stamps from a long-idle bucket
-	// remain in failures until the next NoteFailure fires, and a
-	// caller inspecting bucket size via a future debug surface would
-	// see counts that include already-out-of-window failures.
-	while (!b.failures.empty() && (now - b.failures.front()) > m_cfg.window_seconds) {
-		b.failures.pop_front();
-	}
-	return Decision{};
-}
-
-void CRateLimiter::NoteFailure(const std::string &ip)
-{
-	std::lock_guard<std::mutex> lock(m_mu);
-	const std::time_t now = m_clock();
-	Bucket &b = m_buckets[ip];
-
-	// Sliding window: drop any failure stamp older than
-	// `window_seconds`, then append now. Lockout fires when the live
-	// stamp count crosses `threshold`. Previously the bucket reset
-	// wholesale once `now - window_start > window_seconds`, which
-	// implemented a TUMBLING window — an attacker could split
-	// threshold-1 attempts across the two adjacent windows and
-	// never trip lockout. Per-stamp expiry closes the gap.
-	while (!b.failures.empty() && (now - b.failures.front()) > m_cfg.window_seconds) {
-		b.failures.pop_front();
-	}
-	b.failures.push_back(now);
-
-	if (b.failures.size() >= m_cfg.threshold) {
-		b.lockout_until = now + m_cfg.lockout_seconds;
-	}
-}
-
-void CRateLimiter::NoteSuccess(const std::string &ip)
-{
-	std::lock_guard<std::mutex> lock(m_mu);
-	m_buckets.erase(ip);
-}
-
 // ---------- Header extraction ------------------------------------
 
 std::string ExtractBearerToken(const std::string &authorization_header)
